matching-service/requirements.txt
flask==3.0.0
kafka-python==2.0.2
confluent-kafka==2.3.0
redis==5.0.1
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4
pydantic==2.5.2
python-dotenv==1.0.0
celery==5.3.4
gunicorn==21.2.0
marshmallow==3.20.2
flask-cors==4.0.0
requests==2.31.0
